Described Experiments Added support for float initialization of opinion

+++ code/model/init_parameters.jl
@@ -28,10 +28,10 @@
     βmin::Float64 = 0.7                   # minimum value discount factor
     βmax::Float64 = 1.0                   # maximum value discount factor
     # Determine household sustainability
-    sust_α::Int64 = 1                       # Used in beta distribution for weight of sustainable decisions         <- new
-    sust_β::Int64 = 1                       # Used in beta distribution for weight of sustainable decisions         <- new
-    sust_uncert_α::Int64 = 1                # Used in beta distribution for uncertainty in importance of sustainability     <- new
-    sust_uncert_β::Int64 = 1                # Used in beta distribution for uncertainty in importance of sustainability     <- new
+    sust_α::Float64 = 1.0                       # Used in beta distribution for weight of sustainable decisions         <- new
+    sust_β::Float64 = 1.0                       # Used in beta distribution for weight of sustainable decisions         <- new
+    sust_uncert_α::Float64 = 1.0                # Used in beta distribution for uncertainty in importance of sustainability     <- new
+    sust_uncert_β::Float64 = 1.0                # Used in beta distribution for uncertainty in importance of sustainability     <- new
     sust_upd_rule_politic::Bool = false         # Additional rule for updates, more in notes
     sust_upd_rule_scientific::Bool = false      # Additional rule for updates, more in notes
     sust_upd_rule_use_wealth::Bool = false      # Additional rule for updates, more in notes
